+++ lib/src/database/tables/messages.table.dart
@@ -3,7 +3,7 @@ import 'package:twonly/src/database/tables/contacts.table.dart';
 import 'package:twonly/src/database/tables/groups.table.dart';
 import 'package:twonly/src/database/tables/mediafiles.table.dart';
 
-enum MessageType { media, text, contacts }
+enum MessageType { media, text, contacts, restoreFlameCounter }
 
 @DataClassName('Message')
 class Messages extends Table {