+++ lib/src/views/chats/chat_messages_components/chat_list_entry.dart
@@ -13,6 +13,7 @@ import 'package:twonly/src/services/mediafiles/mediafile.service.dart';
 import 'package:twonly/src/views/chats/chat_messages_components/chat_reaction_row.dart';
 import 'package:twonly/src/views/chats/chat_messages_components/entries/chat_audio_entry.dart';
 import 'package:twonly/src/views/chats/chat_messages_components/entries/chat_contacts.entry.dart';
+import 'package:twonly/src/views/chats/chat_messages_components/entries/chat_flame_restored.entry.dart';
 import 'package:twonly/src/views/chats/chat_messages_components/entries/chat_media_entry.dart';
 import 'package:twonly/src/views/chats/chat_messages_components/entries/chat_text_entry.dart';
 import 'package:twonly/src/views/chats/chat_messages_components/entries/chat_unkown.entry.dart';
@@ -128,6 +129,12 @@ class _ChatListEntryState extends State<ChatListEntry> {
 
     if (widget.message.type == MessageType.contacts.name) {
       return ChatContactsEntry(
+        message: widget.message,
+      );
+    }
+
+    if (widget.message.type == MessageType.restoreFlameCounter.name) {
+      return ChatFlameRestoredEntry(
         message: widget.message,
       );
     }